+++ comics/comics/gunshow.py
@@ -15,6 +15,6 @@ class Crawler(CrawlerBase):
         page_url = 'http://www.gunshowcomic.com/d/%s.html' % (
             pub_date.strftime('%Y%m%d'),)
         page = self.parse_page(page_url)
-        urls = page.src('img[src^="http://www.gunshowcomic.com/comics/"]',
+        urls = page.src('img[src^="http://gunshowcomic.com/comics/"]',
             allow_multiple=True)
         return [CrawlerImage(url) for url in urls]